consolidated CodeOriginRecorder and CodeOriginProbeManager jacoco is happy
